requirements.txt
opencv-python
numpy
tqdm
pillow
imagehash
scikit-image
scipy